Switch from AUC to runtime as meta-target
AUC can be predicted quite well for single base datasets (as done in prior
work), but rather badly for a joint meta-dataset. The problem is that there
are only 38 base datasets, so meta-features describing them might not
generalize well, if you use "fair" splits in meta-cross-validation (all
meta-objects belonging to same base dataset put either into meta-train or
meta-test). Therefore, we switch to runtime as target, which might be more
directly related to the dataset meta-features (and hasn't been analyzed on
this meta-dataset yet).